# public/openapi-tools.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GoldBean API Paid Tools",
    "version": "7.9.5",
    "description": "x402 Pay-per-call API marketplace. Baidu OCR, NLP, TTS, LLM, Image Gen, Search and more.",
    "contact": {
      "url": "https://goldbean-api.xyz"
    }
  },
  "servers": [
    {
      "url": "https://goldbean-api.xyz"
    }
  ],
  "paths": {
...
    "/paid/baidu-ocr": {
      "get": {
        "summary": "Baidu Ocr",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
    "/paid/baidu-ocr-accurate": {
      "get": {
        "summary": "Baidu Ocr Accurate",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
...
    "/paid/baidu-tts": {
      "get": {
        "summary": "Baidu Tts",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
    "/paid/baidu-translate": {
      "get": {
        "summary": "Baidu Translate",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
    "/paid/baidu-llm-chat": {
      "get": {
        "summary": "Baidu Llm Chat",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
...
    "/paid/baidu-image-enhance": {
      "get": {
        "summary": "Baidu Image Enhance",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
...
    "/paid/llm-chat": {
      "get": {
        "summary": "Llm Chat",
        "tags": [
          "paid"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "402": {
            "description": "x402 Payment Required"
          }
        }
      }
    },
...
  }
}